net: dhcpv6: Move DHCPv6 files to lib directory
For consistency with DHCPv4.
