+++ subsys/net/ip/Kconfig.ipv6
@@ -183,12 +183,6 @@ config NET_MAX_6LO_CONTEXTS
 	  6lowpan context options table size. The value depends on your
 	  network and memory consumption. More 6CO options uses more memory.
 
-config NET_DHCPV6
-	bool "DHCPv6 client"
-	select NET_MGMT
-	select NET_MGMT_EVENT
-	depends on NET_UDP
-
 if NET_6LO
 module = NET_6LO
 module-dep = NET_LOG
@@ -214,14 +208,6 @@ module-dep = NET_LOG
 module-str = Log level for IPv6 neighbor cache
 module-help = Enables IPv6 Neighbor Cache code to output debug messages.
 source "subsys/net/Kconfig.template.log_config.net"
-
-if NET_DHCPV6
-module = NET_DHCPV6
-module-dep = NET_LOG
-module-str = Log level for DHCPv6 client
-module-help = Enables DHCPv6 client code to output debug messages.
-source "subsys/net/Kconfig.template.log_config.net"
-endif # NET_DHCPV6
 
 endif # NET_NATIVE_IPV6
 endif # NET_IPV6
