+++ subsys/net/ip/net_core.c
@@ -45,7 +45,7 @@ LOG_MODULE_REGISTER(net_core, CONFIG_NET_CORE_LOG_LEVEL);
 #include "ipv4.h"
 
 #include "dhcpv4/dhcpv4_internal.h"
-#include "dhcpv6_internal.h"
+#include "dhcpv6/dhcpv6_internal.h"
 
 #include "route.h"
