+++ subsys/net/lib/dhcpv6/CMakeLists.txt
@@ -1,0 +1,9 @@
+# Copyright (c) 2024 Nordic Semiconductor ASA
+#
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_library()
+
+zephyr_library_include_directories(${ZEPHYR_BASE}/subsys/net/ip)
+
+zephyr_library_sources_ifdef(CONFIG_NET_DHCPV6 dhcpv6.c)
